refactor: move Cart sub-components to components folder
