+++ pages/cart.js
@@ -1,6 +1,6 @@
-import Content from './components/CartContent/CartContent';
-import CartImage from './components/CartImage/CartImage';
-import CartEmpty from './components/CartEmpty/CartEmpty';
+import Content from '../components/Cart/CartContent/CartContent';
+import CartImage from '../components/Cart/CartImage/CartImage';
+import CartEmpty from '../components/Cart/CartEmpty/CartEmpty';
 import { useSelector } from 'react-redux';
 
 const Index = () => {
